Add `scoredConnections`
This adds `scoredConnections`, which allows you to compute the
`ScoredConnection[]` for a particular node. (The `WeightedGraph` is
required too).

This function is basically the single-node analogue of the full
`decompose` function from `pagerankNodeDecomposition`. I intend for this
function to wholly replace the `decompose` function. Once clients are no
longer calling `decompose`, I will remove that function, rename the
file, and slim down the `ScoredConnection` interface.

Test plan: I duplicated the testing for `decompose` so that we use the
same logic to test `scoredConnections`, and leave the path clear to
remove `decompose` down the line. I verified that the extra snapshot is
identical to the existing snapshot.